Explicit conversion from double to int
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/trunk/FermaNext@243 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ src/TrussUnit/TrussUnitWindow.cpp
@@ -887,9 +887,13 @@ void TrussUnitWindow::createPivotCrossNodes ( QPointArray crossPoints )
         crossNode = *nodeIter;
         // can not allow unfixed node to "eat" a fixed one
         if ( crossNode->getFixation() == Node::Unfixed )
-            node = nodesMergingComparison( *crossNode, 1.5 * nodeFindingPrecision, false );
+            node = nodesMergingComparison( *crossNode, 
+                                           int(1.5 * nodeFindingPrecision),
+                                           false );
         else
-            node = nodesMergingComparison( *crossNode, 1.5 * nodeFindingPrecision, true );
+            node = nodesMergingComparison( *crossNode, 
+                                           int(1.5 * nodeFindingPrecision),
+                                           true );
 
         if ( node )
             mergeNodes ( crossNode, node );
